Ensure same lepton scale +- goes into MET
Previously, as part of runMetCorAndUncFromMiniAOD it would vary the
lepton scale differently to how we want to do our scale variations.
This ensures that the same shift we apply to our lepton goes into
the MET variation.

Side effect: the order of modules changes, meaning that the JER
shift that goes into MET is a bit different as the number of times
the undeterministically-seeded random number generator is called
differs, hence producing differing answers.